+++ bin/remote_payload.js
