+++ lib/remote_payload.coffee
@@ -32,7 +32,7 @@ getExecutionParams = (reqQuery, reqPayload) ->
       catch e
         execParams = reqQuery.params #as handled in platform
   else
-    execParams = _.assign {}, if _.isObject(reqPayload) then reqPayload else {}, reqQuery
+    execParams = _.assign {}, (if _.isObject(reqPayload) then reqPayload else {}), reqQuery
     delete execParams.apikey if execParams.apikey
   execParams
 
@@ -54,14 +54,14 @@ create = (req) ->
       api_key: req.headers['x-cloudmine-apikey'] or req.query.apikey
       app_id: req.params.appid
       session_token: req.headers['x-cloudmine-sessiontoken'] or null
-      user_id: null #snippetConfig.userId -- needs platform request
+      user_id: '[User ID not populated in local deployments]' #snippetConfig.userId -- needs platform request
     params: if _.isEmpty(params) then null else params
     config:
       async: isTruthy req.query.async
       timeout: 30 #getExecTimeout snippetConfig.timeout -- needs platform request
       version: 2 #if snippetConfig.backend_version then parseInt(snippetConfig.backend_version) else DEFAULT_BACKEND_VERSION -- needs platform request
-      type: 'post' 
-    code: undefined 
+      type: 'post'
+    code: undefined
 #  deisRequestBody.response.body = a.platformResult if a.platformResult -- requires platform request
   new_req = {payload: deisRequestBody}
   new_req
