+++ package.json
@@ -2,7 +2,7 @@
   "author": "CloudMine, Inc.",
   "name": "cloudmine-servercode",
   "description": "A Way to run Node.js apps on CloudMine's Internal Platform",
-  "version": "0.9.5",
+  "version": "0.9.6",
   "homepage": "https://cloudmine.me/",
   "license": "MIT",
   "main": "bin/server.js",
